outsource function code to aggregate-energy.ts
